clean up table pipes
