+++ apps/docs/src/app/platform/component-docs/platform-table/platform-table-examples/platform-table-filterable-example.component.html
@@ -35,6 +35,7 @@
         name="description"
         key="description"
         label="Description"
+        [filterable]="true"
     >
     </fdp-column>
 
@@ -43,7 +44,6 @@
         key="price.value"
         label="Price"
         align="end"
-        [filterable]="true"
     >
     </fdp-column>
 
@@ -52,7 +52,6 @@
         key="status"
         label="Status"
         align="center"
-        [filterable]="true"
     ></fdp-column>
 
     <fdp-column
@@ -70,21 +69,26 @@
         column="price"
         label="Price"
         type="custom"
-        [dialogResponsivePadding]="true"
     >
         <ng-container *fdpViewSettingsFilterCustomDef="let model">
-            <label>Minimum Price:</label>
-            <fdp-input
-                type="number"
-                name="min"
-                [(ngModel)]="model.min"
-            ></fdp-input>
-            <label>Maximum Price:</label>
-            <fdp-input
-                type="number"
-                name="max"
-                [(ngModel)]="model.max"
-            ></fdp-input>
+            <fd-layout-panel>
+                <fd-layout-panel-body>
+                    <label>Minimum Price:</label>
+                    <fdp-input
+                        type="number"
+                        name="min"
+                        contentDensity="compact"
+                        [(ngModel)]="model.min"
+                    ></fdp-input>
+                    <label>Maximum Price:</label>
+                    <fdp-input
+                        type="number"
+                        name="max"
+                        contentDensity="compact"
+                        [(ngModel)]="model.max"
+                    ></fdp-input>
+                </fd-layout-panel-body>
+            </fd-layout-panel>
         </ng-container>
     </fdp-table-view-settings-filter>
