+++ libs/platform/src/lib/components/table/pipes/index.ts
@@ -1,3 +1,1 @@
-export * from './sort.pipe';
-export * from './filter.pipe';
 export * from './cell-value.pipe';
